Extract steps into own components
