+++ frontend/src/locales/de.json
@@ -124,6 +124,33 @@
         "moveScheduleEntries": "Aktivitäten verschieben"
       }
     },
+    "campCreate": {
+      "campCreate": {
+        "steps": {
+          "configurate": "Konfigurieren",
+          "infos": "Infos",
+          "template": "Vorlage"
+        }
+      },
+      "campCreateStep1": {
+        "submitTooltip": "Bitte fülle alle Pflichtfelder aus."
+      },
+      "campCreateStep2": {
+        "category": "Block-Kategorie Vorschau",
+        "create": "Lager erstellen",
+        "noContent": "Keine Inhalte",
+        "noPrototype": "Keine Vorlage",
+        "noPrototypeAlert": {
+          "description": "Du musst alle Einstellungen manuell vornehmen. Es sind keine Blockvorlagen & Layouts vorhanden. Dies ist nur für erfahrene Nutzende geeignet.",
+          "title": "Achtung: Du hast \"Keine Vorlage\" ausgewählt."
+        },
+        "progressStates": "Blockstatus Vorschau",
+        "prototypeHint": "Kopiere Kategorien, Blockvorlagen und Blockstatus von einer Lagervorlage",
+        "prototypeHintEmpty": "Keine Kategorien, Blockvorlagen und Blockstatus aus einer Lagervorlage kopieren",
+        "prototypeHintSelected": "Kopiere Kategorien, Blockvorlagen und Blockstatus von dieser Lagervorlage",
+        "submitTooltipPrototype": "Du must noch auswählen ob und welche Lagervorlage du verwenden möchtest."
+      }
+    },
     "collaborator": {
       "collaboratorCreate": {
         "invite": "Einladung verschicken",
@@ -591,27 +618,6 @@
       }
     },
     "campCreate": {
-      "create": "Lager erstellen",
-      "noContent": "Keine Inhalte",
-      "noPrototype": "Keine Vorlage",
-      "noPrototypeAlert": {
-        "description": "Du musst alle Einstellungen manuell vornehmen. Es sind keine Blockvorlagen & Layouts vorhanden. Dies ist nur für erfahrene Nutzende geeignet.",
-        "title": "Achtung: Du hast \"Keine Vorlage\" ausgewählt."
-      },
-      "period": {
-        "add": "Weiteren Lagerabschnitt hinzufügen",
-        "remove": "Diesen Lagerabschnitt entfernen"
-      },
-      "prototypeHint": "Kopiere Kategorien, Blockvorlagen und Blockstatus von einer Lagervorlage",
-      "prototypeHintEmpty": "Keine Kategorien, Blockvorlagen und Blockstatus aus einer Lagervorlage kopieren",
-      "prototypeHintSelected": "Kopiere Kategorien, Blockvorlagen und Blockstatus von dieser Lagervorlage",
-      "steps": {
-        "configurate": "Konfigurieren",
-        "infos": "Infos",
-        "template": "Vorlage"
-      },
-      "submitTooltip": "Bitte fülle alle Pflichtfelder aus.",
-      "submitTooltipPrototype": "Du must noch auswählen ob und welche Lagervorlage du verwenden möchtest.",
       "title": "Lager erstellen"
     },
     "camps": {
